RPi3: improved RTC driver.
It turned out to be not too bright to have date set to 00/00/0000,
as the FAT filesystem driver hated that, and prevented new
files from being created.

RtcEpochSeconds NV variable is used to persist the boot epoch
across boots. If not available, the fixed PcdBootEpochSeconds is
used (set to May 9th, 2018). This should allow you to set
whatever date/time  you want using the Shell date and time
commands.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>