RPi3: "NVRAM" support.
This adds NVRAM support, backed by the RPI_EFI.fd file itself
(the variable store is an additional FV). This is fairly
nice in that the variable store is available before
the FS is discovered, and variables are moved along with
firmware image.

Variable store is persisted to storage on a) reboot b)
ready to boot event.

This is a large change that also take a bump in ATF to
accomodate the larger FD size.

Of course, no persistence in RTS. Hypothetically, if I ever
implement warm reboot, variables created by OS could
be persisted to storage on a warm boot. But that's
quite far away...

The boot timeout and console rows/cols are now persisted.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>