RPi3: allow HypDxe to redirect NT kernel messages to UART
This is opt-in, because it can interfere with kernel
debugging over UART.

We also get module load and unload messages for free
(these don't list bases because I'm lazy).

Also, for 17125 and 17672 (note, this is not a range),
you can override the Kd_Win2000_Mask to improve
verbosity of kernel messages.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>