RPi3: MmcDxe: deal with slow cards
Some bad cards slow down to a crawl in 4-bit mode and
a train of CMD24 writes start timing out. So we ought to
wait until card exits the programming state.

This was already done, but the timeout was arbitrary, and not
coupled to physical time. Make it 1sec (the limits in the SD
simplified spec are much smaller).

Also, refactor since the waiting is done in a few places.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>