RaspberryPiPkg: remove the hardcoded boot options
Once the firmware had no NVRAM support, this was the
only way of ensuring that firmware could be useful
(i.e. prefer external media to just booting the shell).
It never worked too well, because on the Pi 3B+ the USB
hierarchy is different (there's an extra hub).

Now that we have NVRAM, we can dispense with this.
ConnectAll still won't enumerate all the devices in time,
because USB scanning is async because of a lock inversion.

Basically, first time you boot you *will* have to set the
boot order. Then you're fine.

One less edk2Patch, too.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>