RPi3: Fix SPCR definition.
A few dumb bugs means neither Windows nor Linux would know
what to do with the SPCR.

With this change, EMS will work with Windows 10 (obviously,
you need to set 'ems' to on 'on' and 'debug' to 'off' in the
bcd). So you can spawn a cmd shell over serial. Cool!

Update the new release builds as they aren't public yet anyway.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>