RPi3: fix DwUsbDxe license
When I started cleaning this driver up, the original Linaro
code appeared uncomfortably familiar to the U-Boot
implementation, which is why I had documented it as GPL code.
I have done so on the basis of structural and lexical
similarity, having noted just how dissimilar the other other
implementations (reference vendor, FreeBSD, etc) looked in
comparison.

Pete Batard had investigated whether the original Linaro
code was really infringing. He reached out to Marek and
Oleksandr (U-Boot dwc2 driver authors) and they didn't think
there was any blatant Xeroxing going on.

Not sure I fully agree with some of the reasoning, but
if the "victims" don't see plagiarism, then so be it.

Claim this is BSD, not GPL. All of the code written since
the driver was last re-licensed to GPL is mine to re-license.

Signed-off-by: Andrei Warkentin <andrey.warkentin@gmail.com>