Improve timeout handling in DwUsbHostDxe
Networking applications (e.g. iPXE) might experience failures when submitting
a bulk IN for the NIC's RX endpoint, because the bulk IN (correctly) times out
when no received packet is waiting, but DwUsbHostDxe.c treats this as a fatal
error.

With these patches, iPXE is able to successfully download a 128MB test file
via HTTP.

Note: This is a merge of:
- tianocore/edk2-platforms@2b728814e510fa9c9bcd5054bd9fd0db08fcb14e
- tianocore/edk2-platforms@24614a0e84261706e7f70e6fe3fec2a1f78faa89
- tianocore/edk2-platforms@77dc30a13b132e815edafcfa36cbd520ceeb87a9